prevent traversing to parent urls
If not launched on a domain - do not go higher than the folder/ the
crawle is launched on. This can be overriden by specifying -no-parent 0
on the command line

also modifying the cache file logic to use sub folder. with a large and
flat url structure crawling can slow significantly even if all urls are
cached because of the number of files in the same dir