change the default tmp dir location to under /tmp/
having all the tmp files under the application's folder makes it
problematic to use, for example, within cakephp. This is because cake
builds a file map of everything under vendors and therefore ends up
looking in vendors/crawler/tmp/... which will contain millions of files
if you're actaully 'using'this repo.

Therefore, move the tmp dirs to somewhere that isn't going to be indexed
by any normal (php) process